libbinder_ndk: read/write parcelable array am: 90d726a320 am: 5e8d9bc691

+++ libs/binder/ndk/include_ndk/android/binder_parcel.h
@@ -149,7 +149,7 @@ typedef bool (*AParcel_stringArrayElementAllocator)(void* arrayData, size_t inde
  * not required to be null-terminated. If the object at index is null, then this should be null.
  */
 typedef const char* (*AParcel_stringArrayElementGetter)(const void* arrayData, size_t index,
-                                                        int32_t* outLength);
+                                                        size_t* outLength);
 
 /**
  * This is called to allocate an array of size 'length'. If length is -1, then a 'null' array (or
